requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.13.1
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
opencv-python>=4.8.1.78
numpy>=1.26.0
Pillow>=10.1.0
requests>=2.31.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
aiofiles>=23.2.1
insightface>=0.7.3
onnxruntime>=1.16.3

